Move add_derived_field code to FieldInfoContainer.
